requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./bin/update-requirements.sh
#
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools, safety
dparse==0.4.1             # via safety
google-api-python-client==1.7.11  # via -r requirements.in
google-auth-httplib2==0.0.3  # via -r requirements.in, google-api-python-client
google-auth-oauthlib==0.4.1  # via -r requirements.in
google-auth==1.11.2       # via google-api-python-client, google-auth-httplib2, google-auth-oauthlib
httplib2==0.17.0          # via google-api-python-client, google-auth-httplib2
idna==2.9                 # via requests
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.3           # via dparse, safety
pip-tools==4.5.1          # via -r requirements.in
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyparsing==2.4.6          # via packaging
pyyaml==5.3               # via dparse
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via requests-oauthlib, safety
rsa==4.0                  # via google-auth
safety==1.8.5             # via -r requirements.in
six==1.14.0               # via dparse, google-api-python-client, google-auth, packaging, pip-tools
tqdm==4.43.0              # via -r requirements.in
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
